requirements.txt
Django==3.2.9
django-admin-interface==0.17.3
django-cors-headers==3.10.0
django-environ==0.8.1
django-imagefield==0.13.4
django-maintenance-mode==0.16.1
django-reversion==4.0.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.0
PyJWT==2.1.0
PyYAML==6.0
requests==2.26.0
sentry-sdk==1.5.0
whitenoise==5.3.0
